# api/openapi.yaml
openapi: 3.0.3
info:
  title: Keycast HTTP API
  version: 1.0.0
  description: |
    # Keycast HTTP Signing API

    Fast HTTP signing for Nostr events without NIP-46 relay overhead. Keycast provides both traditional NIP-46 remote signing and a high-performance HTTP API for low-latency signing operations.

    ## Authentication

    Most endpoints require a UCAN token obtained from the `/auth/login` or `/auth/register` endpoints. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Performance

    - **Fast Path**: ~10ms signing using cached handlers (unified mode)
    - **Slow Path**: ~50-100ms signing using database + KMS decryption

    ## Security Notes

    ⚠️ **Current Limitations**:
    - No permission validation on HTTP signing endpoint
    - No rate limiting (endpoints can be called unlimited times)
    - UCAN tokens valid for 24 hours

    These limitations are being addressed in upcoming releases.

  contact:
    name: Keycast Support
    url: https://github.com/rabble/keycast
  license:
    name: MIT
servers:
  - url: https://entryway.divine.video/api
    description: Production ATProto Authorization Server
  - url: https://login.divine.video/api
    description: Production Divine and Nostr login server
  - url: http://localhost:3000/api
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: UCAN token from /auth/register or /auth/login
    ServiceToken:
      type: http
      scheme: bearer
      description: KEYCAST_SERVICE_TOKEN for trusted service-to-service callers (e.g. the AP gateway)
...
paths:
  /auth/register:
    post:
      summary: Register new user
      description: |
        Create a new user account with email/password authentication. Automatically generates a Nostr keypair and returns a UCAN token for immediate use.

        **BYOK Support**: You can optionally provide your own Nostr secret key via the `nsec` parameter. If not provided, a new key will be generated.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                summary: Register with auto-generated key
                value:
                  email: "user@example.com"
                  password: "SecurePassword123"
              byok:
                summary: Register with your own key (BYOK)
                value:
                  email: "user@example.com"
                  password: "SecurePassword123"
                  nsec: "nsec1abc123..."
      responses:
        '200':
          description: Successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request (weak password, invalid nsec)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login
      description: Authenticate with email/password and receive a UCAN token valid for 24 hours.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/sign:
    post:
      summary: Sign Nostr event (Fast HTTP signing)
      description: |
        Sign a Nostr event via HTTP - **10-50x faster than NIP-46**.

        ## Performance
        - **Fast path** (~10ms): Uses cached authorization handlers in unified mode
        - **Slow path** (~50-100ms): Decrypts key from database using KMS

        ## ⚠️ Security Warning
        **This endpoint currently has NO permission validation**. Any authenticated user can sign any event kind with any content. Permission checks are coming in a future release.

        ## Supported Event Kinds
        Currently accepts all Nostr event kinds (0-65535), including:
        - `kind: 0` - User metadata
        - `kind: 1` - Short text note
        - `kind: 3` - Contact list
        - `kind: 4` - Encrypted direct message
        - `kind: 5` - Deletion
        - `kind: 1984` - Reporting
        - `kind: 9734-9735` - Zaps
        - `kind: 10000+` - Replaceable events
        - `kind: 20000+` - Ephemeral events
        - `kind: 30000+` - Parameterized replaceable events
      tags:
        - Signing
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignEventRequest'
            examples:
              text_note:
                summary: Sign a text note (kind 1)
                value:
                  event:
                    kind: 1
                    content: "Hello from Keycast!"
                    tags: []
                    created_at: 1234567890
              dm:
                summary: Sign encrypted DM (kind 4)
                value:
                  event:
                    kind: 4
                    content: "encrypted_content_here"
                    tags: [["p", "recipient_pubkey"]]
                    created_at: 1234567890
      responses:
        '200':
          description: Event successfully signed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignEventResponse'
        '401':
          description: Unauthorized (invalid or missing JWT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found or key not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Signing failed (KMS error, invalid key, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ap/keys:
    post:
      summary: Create or return an actor's ActivityPub RSA key
      description: |
        Mints (or returns, idempotently) an RSA-2048 keypair for an actor and returns its
        SPKI public PEM for embedding in `publicKey.publicKeyPem`. The private key is stored
        encrypted-at-rest and is never returned. Calling again returns the existing key
        with `created: false` — the key is never regenerated (federated keys are cached by
        remote servers). Auth: ServiceToken (with `pubkey` or `actor`) or an OAuth UCAN
        for the same actor.
      tags:
        - ActivityPub
      security:
        - ServiceToken: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApCreateKeyRequest'
      responses:
        '200':
          description: Key created or already existed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApKeyResponse'
        '400':
          description: Invalid or missing pubkey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: UCAN pubkey mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ap/keys/{actor_or_pubkey}:
    get:
      summary: Get an actor's ActivityPub public key PEM
      description: |
        Returns the actor's public PEM, lazily minting the RSA key if needed. The path
        accepts either a hex Nostr pubkey or a URL-encoded ActivityPub actor URL.
      tags:
        - ActivityPub
      security:
        - ServiceToken: []
        - BearerAuth: []
      parameters:
        - in: path
          name: actor_or_pubkey
          required: true
          schema:
            type: string
          description: Hex Nostr pubkey or URL-encoded ActivityPub actor URL
      responses:
        '200':
          description: Public key PEM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApPublicKeyResponse'
        '404':
          description: No AP key for this actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ap/sign:
    post:
      summary: Sign an HTTP-Signature signing string (RSA-SHA256)
      description: |
        Signs the exact caller-supplied signing string with RSASSA-PKCS1-v1_5 + SHA-256 and
        returns the base64 signature. The caller builds the draft-cavage signing string and
        assembles the final `Signature:` header. The private key never leaves Keycast.
        Suspended/banned accounts are rejected. Service-token callers identify the actor
        with `pubkey` or `actor`; UCAN callers need a live OAuth authorization for the
        same actor.
      tags:
        - ActivityPub
      security:
        - ServiceToken: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApSignRequest'
      responses:
        '200':
          description: Signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApSignResponse'
        '403':
          description: Account restricted or pubkey mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No AP key for this actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/bunker:
    get:
      summary: Get NIP-46 bunker URL
      description: |
        Retrieve the NIP-46 bunker connection string for traditional relay-based signing.

        Use this when you need full NIP-46 compatibility with permission validation, or when HTTP signing is not available.
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bunker URL retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BunkerUrlResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/pubkey:
    get:
      summary: Get user's public key
      description: Returns the user's Nostr public key in both hex and npub (bech32) formats.
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Public key retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubkeyResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/profile:
    get:
      summary: Get user profile
      description: Retrieve the authenticated user's profile information including email, username, and Nostr metadata.
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Update user profile
      description: Update username, display name, or avatar URL for the authenticated user.
      tags:
        - User
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "alice"
                display_name:
                  type: string
                  example: "Alice Smith"
                avatar_url:
                  type: string
                  format: uri
                  example: "https://example.com/avatar.jpg"
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/atproto/enable:
    post:
      summary: Enable ATProto provisioning
      description: Enable ATProto opt-in after username claim, return the claimed username with a pending lifecycle state, and trigger provisioning.
      tags:
        - User
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtprotoEnableRequest'
      responses:
        '202':
          description: ATProto provisioning accepted and lifecycle set to pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtprotoStatusResponse'
        '400':
          description: Username missing, unclaimed, or mismatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/atproto/status:
    get:
      summary: Get ATProto lifecycle status
      description: Returns the authenticated user's enabled flag, lifecycle state, DID, last provisioning error, and claimed username.
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current ATProto status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtprotoStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/atproto/disable:
    post:
      summary: Disable ATProto publishing
      description: Disable ATProto publishing for the authenticated user, trigger downstream cleanup, and return enabled=false with state=disabled.
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ATProto publishing disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtprotoStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/sessions:
    get:
      summary: List active sessions
      description: Get all active NIP-46 authorization sessions for the authenticated user.
      tags:
        - User
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sessions retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...